packages/cli/src/commands/svg.ts
import { readFileSync, mkdirSync, writeFileSync } from "node:fs";
import { resolve as resolvePath, join as joinPath, dirname } from "path";
import glob from "glob";
// @ts-expect-error
import SVGSpriter from "svg-sprite";
import File from "vinyl";
import { hasHelpFlag } from "../utils.js";

export const svgHelpMessage = `
  Atomic Smash CLI - SVG command.

  Generate an SVG sprite from a group of SVGs.

  Options:
    --in         The directory where the SVGs can be found.
    --out        The directory where the SVG sprite will be output.

  Example usage:
    $ smash-cli svg --in icons --out public/assets
`;

export default function svg(args: string[]) {
	if (hasHelpFlag(args)) {
		console.log(svgHelpMessage);
		return;
	}
	const inFlag = args[args.findIndex((arg) => arg === "--in") + 1];
	if (!inFlag || inFlag.startsWith("--")) {
		throw new Error("You need to provide a value for the --in flag.");
	}
	const outFlag = args[args.findIndex((arg) => arg === "--out") + 1];
	if (!outFlag || outFlag.startsWith("--")) {
		throw new Error("You need to provide a value for the --out flag.");
	}

	const config = {
		dest: outFlag,
		shape: {
			dimension: {
				attributes: false,
			},
			transform: [
				{
					svgo: {},
				},
			],
		},
		mode: {
			symbol: {
				dest: "",
				sprite: "sprite.svg",
			},
		},
	};
	const spriter = new SVGSpriter(config);

	const cwd = resolvePath(inFlag);
	// Find SVG files recursively via `glob`
	glob("**/*.svg", { cwd }, async (error, files) => {
		if (error) {
			console.error(error.message);
			throw error;
		}
		console.log({ files });
		await new Promise<void>((resolve, reject) => {
			files.forEach((file) => {
				// Create and add a vinyl file instance for each SVG
				spriter.add(
					new File({
						path: joinPath(cwd, file), // Absolute path to the SVG file
						base: cwd, // Base path (see `name` argument)
						contents: readFileSync(joinPath(cwd, file)), // SVG file contents
					})
				);
			});
			resolve();
		})
			.then(async () => {
				await spriter
					.compileAsync()
					.then((compiledResponse: any) => {
						const { result } = compiledResponse;
						for (const type of Object.values<any>(result.symbol)) {
							mkdirSync(dirname(type.path), { recursive: true });
							writeFileSync(type.path, type.contents);
						}
					})
					.catch((error: any) => {
						console.error({ error });
						throw error;
					});
			})
			.then(() => {
				console.log(`SVG sprite was successfully generated in ${outFlag}.`);
			})
			.catch((error) => {
				console.error({ error });
				throw error;
			});
	});
}
